requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
async-timeout==4.0.3
attrs==23.2.0
backoff==1.11.1
blurhash==1.1.4
certifi==2024.2.2
charset-normalizer==3.3.2
decorator==5.1.1
frozenlist==1.4.1
idna==3.6
Mastodon.py==1.8.1
multidict==6.0.5
python-dateutil==2.9.0.post0
python-magic==0.4.27
redis==5.0.2
requests==2.31.0
six==1.16.0
spotify==0.10.2
spotipy==2.23.0
urllib3==2.2.1
yarl==1.9.4
